processor/requirements.txt
redis
psycopg2-binary
pydantic
